// binders-devops-service-v1/app/src/dc/docker-compose-spec.d.ts
export type ListOfStrings = string[];
// ...
export type ListOrDict =
    | {
          /**
           * This interface was referenced by `undefined`'s JSON-Schema definition
           * via the `patternProperty` ".+".
           */
          [k: string]: string | number | boolean | null;
      }
    | string[];
// ...
export type ExtraHosts =
    | {
          /**
           * This interface was referenced by `undefined`'s JSON-Schema definition
           * via the `patternProperty` ".+".
           */
          [k: string]: string | string[];
      }
    | string[];
export type ServiceConfigOrSecret = (
    | string
    | {
          source?: string;
          target?: string;
          uid?: string;
          gid?: string;
          mode?: number | string;
          /**
           * This interface was referenced by `undefined`'s JSON-Schema definition
           * via the `patternProperty` "^x-".
           */
          [k: string]: unknown;
      }
)[];
// ...
export interface Service {
    develop?: Development;
    deploy?: Deployment;
    annotations?: ListOrDict;
    attach?: boolean | string;
    build?:
        | string
        | {
              context?: string;
              dockerfile?: string;
              dockerfile_inline?: string;
              entitlements?: string[];
              args?: ListOrDict;
              ssh?: ListOrDict;
              labels?: ListOrDict;
              cache_from?: string[];
              cache_to?: string[];
              no_cache?: boolean | string;
              additional_contexts?: ListOrDict;
              network?: string;
              pull?: boolean | string;
              target?: string;
              shm_size?: number | string;
              extra_hosts?: ExtraHosts;
              isolation?: string;
              privileged?: boolean | string;
              secrets?: ServiceConfigOrSecret;
              tags?: string[];
              ulimits?: Ulimits;
              platforms?: string[];
              /**
               * This interface was referenced by `undefined`'s JSON-Schema definition
               * via the `patternProperty` "^x-".
               */
              [k: string]: unknown;
          };
// ...
    depends_on?:
        | ListOfStrings
        | {
              /**
               * This interface was referenced by `undefined`'s JSON-Schema definition
               * via the `patternProperty` "^[a-zA-Z0-9._-]+$".
               */
              [k: string]: {
                  restart?: boolean | string;
                  required?: boolean;
                  condition: "service_started" | "service_healthy" | "service_completed_successfully";
                  /**
                   * This interface was referenced by `undefined`'s JSON-Schema definition
                   * via the `patternProperty` "^x-".
                   */
                  [k: string]: unknown;
              };
          };
// ...
}
export interface ServiceHook {
    command: Command;
    user?: string;
    privileged?: boolean | string;
    working_dir?: string;
    environment?: ListOrDict;
    /**
     * This interface was referenced by `ServiceHook`'s JSON-Schema definition
     * via the `patternProperty` "^x-".
     */
    [k: string]: unknown;
}
export interface Ulimits {
    /**
     * This interface was referenced by `Ulimits`'s JSON-Schema definition
     * via the `patternProperty` "^[a-z]+$".
     */
    [k: string]:
        | (number | string)
        | {
              hard: number | string;
              soft: number | string;
              /**
               * This interface was referenced by `undefined`'s JSON-Schema definition
               * via the `patternProperty` "^x-".
               */
              [k: string]: unknown;
          };
}